+++ fn-extra/CHANGELOG.md
@@ -3,6 +3,14 @@
   - Switch Heist to use StateT rather than ReaderT, so splices can
     update context.
   - Add support for compiled heist.
+  - Add support for digestive functors. Note: current (significant)
+    limitation is that if you have _already_ parsed the request body
+    with `!=>` then we can only use post params and not files. This is
+    because `Fn` parses them to lazy bytestrings, but digestive
+    functors wants them in temporary files, and for now, we're not
+    converting (it could be done). But, the expectation is that if you
+    are using digestive functors, you won't ever be parsing post
+    bodies with `!=>`.
 
 * 0.2.0.1 Daniel Patterson <dbp@dbpmail.net> 2016-1-20